+++ map-widget/qml/MapWidget.qml
