+++ map-widget/qml/MapWidgetContextMenu.qml
