+++ map-widget/qml/MapWidgetError.qml
