+++ map-widget/qml/icons/mapwidget-context-menu.png
