+++ map-widget/qml/icons/mapwidget-marker-gray.png
