+++ map-widget/qml/icons/mapwidget-marker-selected.png
