+++ map-widget/qml/icons/mapwidget-marker.png
