+++ map-widget/qml/icons/mapwidget-toggle-satellite.png
