+++ map-widget/qml/icons/mapwidget-toggle-street.png
