+++ map-widget/qml/icons/mapwidget-zoom-in.png
