+++ map-widget/qml/icons/mapwidget-zoom-out.png
