+++ map-widget/qmlmapwidgethelper.cpp
