+++ map-widget/qmlmapwidgethelper.h
