+++ subsurface.qrc
@@ -1,16 +1,16 @@
 <RCC>
     <qresource prefix="/">
-        <file alias="MapWidget.qml">mobile-widgets/qml/MapWidget.qml</file>
-        <file alias="MapWidgetError.qml">mobile-widgets/qml/MapWidgetError.qml</file>
-        <file alias="MapWidgetContextMenu.qml">mobile-widgets/qml/MapWidgetContextMenu.qml</file>
-        <file alias="mapwidget-marker">mobile-widgets/qml/icons/mapwidget-marker.png</file>
-        <file alias="mapwidget-marker-gray">mobile-widgets/qml/icons/mapwidget-marker-gray.png</file>
-        <file alias="mapwidget-marker-selected">mobile-widgets/qml/icons/mapwidget-marker-selected.png</file>
-        <file alias="mapwidget-toggle-satellite">mobile-widgets/qml/icons/mapwidget-toggle-satellite.png</file>
-        <file alias="mapwidget-toggle-street">mobile-widgets/qml/icons/mapwidget-toggle-street.png</file>
-        <file alias="mapwidget-context-menu">mobile-widgets/qml/icons/mapwidget-context-menu.png</file>
-        <file alias="mapwidget-zoom-in">mobile-widgets/qml/icons/mapwidget-zoom-in.png</file>
-        <file alias="mapwidget-zoom-out">mobile-widgets/qml/icons/mapwidget-zoom-out.png</file>
+        <file alias="MapWidget.qml">map-widget/qml/MapWidget.qml</file>
+        <file alias="MapWidgetError.qml">map-widget/qml/MapWidgetError.qml</file>
+        <file alias="MapWidgetContextMenu.qml">map-widget/qml/MapWidgetContextMenu.qml</file>
+        <file alias="mapwidget-marker">map-widget/qml/icons/mapwidget-marker.png</file>
+        <file alias="mapwidget-marker-gray">map-widget/qml/icons/mapwidget-marker-gray.png</file>
+        <file alias="mapwidget-marker-selected">map-widget/qml/icons/mapwidget-marker-selected.png</file>
+        <file alias="mapwidget-toggle-satellite">map-widget/qml/icons/mapwidget-toggle-satellite.png</file>
+        <file alias="mapwidget-toggle-street">map-widget/qml/icons/mapwidget-toggle-street.png</file>
+        <file alias="mapwidget-context-menu">map-widget/qml/icons/mapwidget-context-menu.png</file>
+        <file alias="mapwidget-zoom-in">map-widget/qml/icons/mapwidget-zoom-in.png</file>
+        <file alias="mapwidget-zoom-out">map-widget/qml/icons/mapwidget-zoom-out.png</file>
         <file alias="satellite.svg">icons/satellite.svg</file>
         <file alias="star">icons/star.svg</file>
         <file alias="subsurface-icon">icons/subsurface-icon.png</file>
